add tdd testing in usecase

+++ src/main/kotlin/com/example/dddtesting/ipc/application/response/IpcYearValueResponse.kt
@@ -2,4 +2,4 @@ package com.example.dddtesting.ipc.application.response
 
 import java.time.Year
 
-data class IpcYearValueResponse(val year: Year, val value: Int)
+data class IpcYearValueResponse(val year: Year, val value: Double)
