+++ src/main/kotlin/com/example/dddtesting/ipc/domain/entity/IpcDataEntity.kt
@@ -2,4 +2,4 @@ package com.example.dddtesting.ipc.domain.entity
 
 import java.util.*
 
-data class IpcDataEntity(private val month: Date, private val value: Int)
+data class IpcDataEntity(private val month: Date, private val value: Double)
